+++ R/table_getNearestDistance.R
@@ -8,6 +8,8 @@
 #' @param longitude Vector of longitudes in decimal degrees E.
 #' @param latitude Vector of latitudes in decimal degrees N.
 #' @param radius Radius in meters.
+#' @param measure One of "haversine" "vincenty", "geodesic", or "cheap" 
+#' specifying desired method of geodesic distance calculation. See \code{?geodist::geodist}.
 #' @return Vector of distances from known locations.
 #' @examples
 #' library(MazamaLocationUtils)
@@ -32,7 +34,8 @@ table_getNearestDistance <- function(
   locationTbl = NULL,
   longitude = NULL,
   latitude = NULL,
-  radius = NULL
+  radius = NULL,
+  measure = "geodesic"
 ) {
 
   # ----- Validate parameters --------------------------------------------------
@@ -61,7 +64,7 @@ table_getNearestDistance <- function(
       paired = FALSE,
       sequential = FALSE,
       pad = FALSE,
-      measure = "geodesic"
+      measure = measure
     )
   
   # NOTE:  distance matrix is nrow(locationTbl) X length(longitude)