+++ tests/testthat/test-table_findOverlappingDistances.R
@@ -1,16 +1,16 @@
-context("table_findOverlappingLocations")
+context("table_findOverlappingDistances")
 
 test_that("bad/missing input generates an error", {
   
   # tbl not a dataframe
-  expect_error(table_findOverlappingLocations(tbl <- c("vector", "instead"), radius = 10))
+  expect_error(table_findOverlappingDistances(locationTbl = c("vector", "instead"), radius = 10))
   # radius not numeric
-  expect_error(table_findOverlappingLocations(tbl <- wa_airfire_meta, radius = "bad radius"))
+  expect_error(table_findOverlappingDistances(locationTbl = wa_airfire_meta, radius = "bad radius"))
   
   # tbl missing
-  expect_error(table_findOverlappingLocations(tbl <-NULL, radius = 10))
+  expect_error(table_findOverlappingDistances(tbl <-NULL, radius = 10))
   # radius missing
-  expect_error(table_findOverlappingLocations(tbl <- wa_airfire_meta, radius = NULL))
+  expect_error(table_findOverlappingDistances(locationTbl = wa_airfire_meta, radius = NULL))
   
 })
 
@@ -18,17 +18,17 @@ test_that("table missing latitude or longitude generates an error", {
   
   # No latitude
   testTbl <- wa_airfire_meta %>% dplyr::select(-latitude)
-  expect_error(table_findOverlappingLocations(tbl <- testTbl, radius = 10))
+  expect_error(table_findOverlappingDistances(locationTbl = testTbl, radius = 10))
   
   # No longitude
   testTbl <- wa_airfire_meta %>% dplyr::select(-longitude)
-  expect_error(table_findOverlappingLocations(tbl <- testTbl, radius = 10))
+  expect_error(table_findOverlappingDistances(locationTbl = testTbl, radius = 10))
   
 })
 
 test_that("passing radius = 1000 returns empty tibble", {
   
-  result <- table_findOverlappingLocations(wa_airfire_meta, radius = 1000)
+  result <- table_findOverlappingDistances(wa_airfire_meta, radius = 1000)
   
   # Is it the right type?
   expect_equal(class(result), c("tbl_df", "tbl", "data.frame"))
@@ -40,7 +40,7 @@ test_that("passing radius = 1000 returns empty tibble", {
 
 test_that("passing radius = 2000 returns expected output", {
   
-  result <- table_findOverlappingLocations(wa_airfire_meta, radius = 2000)
+  result <- table_findOverlappingDistances(wa_airfire_meta, radius = 2000)
   
   # Is it the right type?
   expect_equal(class(result), c("tbl_df", "tbl", "data.frame"))